services/data/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pymongo
redis
tenacity
